Tartibga keltirish (modul tartibi)

+++ lms_kod_topshiriqlari/03-Shartlar_va_Sikllar/03-__Shartli_menyu___oddiy_CLI_menyu/05-5__Menyu__kalkulyator__faqat____-_.py
@@ -1,6 +1,6 @@
 a, b, op = input().split()
-a = float(a)
-b = float(b)
+a = int(a)
+b = int(b)
 
 if op == "+":
     print(a + b)
